* master:
   TWinSCPFileSystem::ImportSessions: cleanup
   TXmlStorage::Readbool: fix
   ChangeLog: update
   ChangeLog: update
   Classes.h: fix x64-related warnings
   Common: StrToFloat, StrToFloatDef, Int64ToStr
  TXmlStorage: implement import from .netbox file
  TXmlStorage: implement export to .netbox file
   TWinSCPFileSystem::ExportSession
   add THierarchicalStorage::Init
  move TXmlStorage to XmlStorage.cpp
   TWinSCPFileSystem::ImportSessions: use TXmlStorage
  TWinSCPFileSystem::ExportSession: use TXmlStorage
  add class TXmlStorage
   enum TStorage { stRegistry, stXmlFile };
   CopyParam.cpp: fix x64-related warnings
  TCopyParamType::ValidLocalFileName, TCopyParamType::ReplaceChar: fix
  CHERRY-PICK: Common: StrToFloat, StrToFloatDef, Int64ToStr; Classes.h: fix x64-related warnings
   NetBox.vcxproj: add ..\..\libs\tinyXML

Conflicts:
	src/NetBox/ChangeLog
	src/NetBox/NetBox.vcxproj
	src/NetBox/WinSCPFileSystem.cpp
	src/core/HierarchicalStorage.h

+++ src/core/Classes.h
@@ -33,7 +33,7 @@ std::string W2MB(const wchar_t *src, const UINT cp = CP_ACP);
 inline int __cdecl debug_printf(const wchar_t *format, ...)
 {
     (void)format;
-    int len = 0;
+    size_t len = 0;
 #ifdef NETBOX_DEBUG
     va_list args;
     va_start(args, format);
@@ -50,7 +50,7 @@ inline int __cdecl debug_printf(const wchar_t *format, ...)
 inline int __cdecl debug_printf2(const char *format, ...)
 {
     (void)format;
-    int len = 0;
+    size_t len = 0;
 #ifdef NETBOX_DEBUG
     va_list args;
     va_start(args, format);
